Fixed search query!

+++ server/src/api/search/models/search.rs
@@ -24,7 +24,7 @@ pub struct SearchPost {
     pub url : Option<String>,
     pub name : String,
     pub body : Option<String>,
-    pub score : i64,
+    pub score : i32,
     pub actor_id : String,
     pub community_name : String,
     pub comments : Vec<SearchComment>
@@ -33,7 +33,7 @@ pub struct SearchPost {
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct SearchComment {
     pub content : String,
-    pub score : i64
+    pub score : i32
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
